refactor: move images and fonts to assets folder
