--- src/components/login-provider-button/index.tsx
@@ -1,3 +1,0 @@
-import LoginProviderButton from './login-provider-button'
-
-export default LoginProviderButton
