--- src/components/navigation/index.tsx
@@ -1,3 +1,0 @@
-import Navigation from './navigation'
-
-export default Navigation
